Compute v2: Suppress Availability Zone Diff, when an old AZ is empty
